+++ src/components/Home/ProfilePicture.tsx
@@ -68,9 +68,9 @@ const ProfilePicture = ({
 
       <Animated.View
         style={tw`z-10 absolute rounded-3xl inset-0.5 bg-gray-300 dark:bg-gray-700 overflow-hidden`}>
-        {pending ? (
-          <LoadingSkeleton height={`100%`} width={`100%`} />
-        ) : url ? (
+        {pending ? <LoadingSkeleton height={`100%`} width={`100%`} /> : null}
+
+        {url ? (
           <Image
             cachePolicy="memory-disk"
             contentFit="cover"
@@ -78,13 +78,15 @@ const ProfilePicture = ({
               uri: url,
               cacheKey: `${url}-${dayjs().format('YYYY-MM-DD')}`,
             }}
-            style={tw`h-full w-full`}
+            style={tw`absolute h-full w-full z-10`}
           />
-        ) : initials ? (
+        ) : null}
+
+        {initials ? (
           <View
             style={[
               tw`h-full w-full flex items-center justify-center`,
-              {
+              !url && {
                 backgroundColor: getColorFromSeed(initials),
               },
             ]}>
@@ -93,9 +95,9 @@ const ProfilePicture = ({
               style={[
                 tw`text-xl font-bold self-center m-auto`,
                 initialsStyle,
-                {
-                  color: invertColor(getColorFromSeed(initials), true),
-                },
+                !url
+                  ? { color: invertColor(getColorFromSeed(initials), true) }
+                  : tw`text-gray-700 dark:text-gray-300`,
               ]}>
               {initials}
             </AppText>
@@ -109,7 +111,7 @@ const ProfilePicture = ({
         <Animated.View
           entering={BounceIn.duration(1000).delay(300)}
           exiting={BounceOut.duration(1000)}
-          style={tw`z-10 h-5 w-5 bg-gray-100 dark:bg-black rounded-full absolute flex items-center justify-center -bottom-0.5 -right-0.5`}>
+          style={tw`z-20 h-5 w-5 bg-gray-100 dark:bg-black rounded-full absolute flex items-center justify-center -bottom-0.5 -right-0.5`}>
           <View style={tw`h-3 w-3 bg-emerald-600 dark:bg-emerald-700 rounded-full`} />
         </Animated.View>
       )}